refactor(homepage): extract donation components and tests
